[experiment] Metro GPU Hu moments experiment - double to long.

+++ modules/gpu-processing/include/cir/gpuprocessing/count_moments.cuh
@@ -11,7 +11,7 @@ double count_raw_moment(uchar* data, int width, int height, int step, int p, int
 		cudaStream_t stream);
 
 __global__
-void k_count_raw_moment(uchar* data, int width, int height, int step, int p, int q, double* result);
+void k_count_raw_moment(uchar* data, int width, int height, int step, int p, int q, long* result);
 
 __device__
 int pow(int p, int q);
